backend/requirements.txt
## Setup py requirements
requests==2.22.0
boto3==1.9.146
matplotlib==3.0.3
simplejson==3.16.0
attrs==19.1.0
jsonpickle==1.1
redis==3.2.1
fakeredis==1.0.3
pymongo==3.8.0
numpy==1.16.3
scipy==1.2.1
tensorflow==1.13.1
apscheduler==3.6.0
tqdm==4.31.0
networkx==2.3
fbmessenger==6.0.0
pykwalify==1.7.0
coloredlogs==10.0
ruamel.yaml==0.15.94
scikit-learn==0.20.2
slackclient==1.3.1
python-telegram-bot==11.1.0
twilio==6.26.3
webexteamssdk==1.1.1
mattermostwrapper==2.1
rocketchat_API==0.6.31
colorhash==1.0.2
pika==1.0.1
jsonschema==2.6.0
packaging==19.0
gevent==1.4.0
pytz==2019.1
python-dateutil==2.8.0
rasa-sdk~=1.1.0
colorclass==2.2.0
terminaltables==3.1.0
sanic==19.3.1
sanic-cors==0.9.8
sanic-jwt==1.3.1
aiohttp==3.5.4
questionary==1.1.1
python-socketio==4.0.1
pydot==1.4.1
async_generator==1.10
SQLAlchemy~=1.3.3
kafka-python==1.4.6
sklearn-crfsuite==0.3.6
psycopg2-binary==2.8.2
pandas~=0.24.1
pydictionary
nltk==3.4.4
wiktionaryparser==0.0.97
beautifulsoup4==4.4.0
deepdiff==3.3.0
